+++ resources/core/cljs/home.cljs
@@ -6,6 +6,8 @@
 (ns <<namespace>>.client.home
   (:require [re-frame.core :as rf]<<#hydrogen-session?>>
             [<<namespace>>.client.session :as session]
+            [<<namespace>>.client.tooltip :as tooltip]
+            [<<namespace>>.client.tooltip.loading-popup :as loading-popup]
             [<<namespace>>.client.user :as user]<</hydrogen-session?>>
             [<<namespace>>.client.view :as view]))
 
@@ -14,6 +16,32 @@
  (fn [_ _]
    {:dispatch [::view/set-active-view :home]
     :redirect "/#/home"}))<<#hydrogen-session?>>
+
+;; This section contains code solely for demo purposes.
+;; Go ahead and delete it.
+(rf/reg-event-fx
+  ::DELETEME-demo-request
+  (fn [_ _]
+      {:dispatch [::loading-popup/set-loading "Faking loading. I'll disappear after 5 seconds..."]
+       ;; This http-xhrio should be there instead of the dispatch-later in real situation.
+       ; http-xhrio {...}
+       :dispatch-later [{:ms 5000
+                         :dispatch [::loading-popup/stop-loading]}]}))
+
+(defn- DELETEME-demo-tooltip-component []
+       [:div
+        "Hello! I'm a demo tooltip."])
+
+(defn- DELETEME-demo-tooltip []
+  [:button.btn
+   {:on-click #(rf/dispatch [::tooltip/register {:id "generic-popup"
+                                                 :component DELETEME-demo-tooltip-component}])}
+   "Demo tooltip"])
+
+(defn- DELETEME-demo-loading []
+  [:button.btn {:on-click #(rf/dispatch [::DELETEME-demo-request])}
+   "Pretend you are requesting something from the Internet"])
+;; end of obsolete demo code
 
 (defn- user-details []
   (let [user-data (rf/subscribe [::user/user-data])]
@@ -33,6 +61,8 @@
 (defn main []
   [:div {:id "home"}
    [:img {:src "images/hydrogen-logo-white.svg" :alt "Hydrogen logo"}]
-   [:h1 "Welcome to Hydrogen!"]<<#hydrogen-session?>>
+   [:h1 "Welcome to Hydrogen!"]
+   [DELETEME-demo-tooltip]
+   [DELETEME-demo-loading]<<#hydrogen-session?>>
    [user-details]
    [logout]<</hydrogen-session?>>])